Add multiple authors support to Blog section
This update includes a change from singular to multiple authors support in the Blog section of the application. Changes have been made to the Blog component, the UserCardImage component, and the 'authorData.ts' and 'model.ts' data models to reflect this new feature. Now more than one author can be displayed and sorted in the authors section.