Add author search function in blog section
A search bar has been added to the Authors section of the Blog page. It allows users to filter the author list by typing in the search bar, providing a more convenient and user-friendly browsing experience. Changes include both UI CSS updates and the introduction of a state for the filtered Authors list.